refactor(core): move public extensions to specific package
