+++ core/src/main/java/io/github/thibaultbee/streampack/core/internal/sources/video/camera/CameraSettings.kt
@@ -35,21 +35,21 @@ import io.github.thibaultbee.streampack.core.internal.utils.extensions.normalize
 import io.github.thibaultbee.streampack.core.internal.utils.extensions.rotate
 import io.github.thibaultbee.streampack.core.logger.Logger
 import io.github.thibaultbee.streampack.core.streamers.DefaultCameraStreamer
-import io.github.thibaultbee.streampack.core.utils.getAutoExposureModes
-import io.github.thibaultbee.streampack.core.utils.getAutoFocusModes
-import io.github.thibaultbee.streampack.core.utils.getAutoWhiteBalanceModes
-import io.github.thibaultbee.streampack.core.utils.getCameraCharacteristics
-import io.github.thibaultbee.streampack.core.utils.getExposureMaxMeteringRegionsSupported
-import io.github.thibaultbee.streampack.core.utils.getExposureRange
-import io.github.thibaultbee.streampack.core.utils.getExposureStep
-import io.github.thibaultbee.streampack.core.utils.getFocusMaxMeteringRegionsSupported
-import io.github.thibaultbee.streampack.core.utils.getLensDistanceRange
-import io.github.thibaultbee.streampack.core.utils.getScalerMaxZoom
-import io.github.thibaultbee.streampack.core.utils.getSensitivityRange
-import io.github.thibaultbee.streampack.core.utils.getWhiteBalanceMeteringRegionsSupported
-import io.github.thibaultbee.streampack.core.utils.getZoomRatioRange
-import io.github.thibaultbee.streampack.core.utils.isFlashAvailable
-import io.github.thibaultbee.streampack.core.utils.isOpticalStabilizationAvailable
+import io.github.thibaultbee.streampack.core.utils.extensions.getAutoExposureModes
+import io.github.thibaultbee.streampack.core.utils.extensions.getAutoFocusModes
+import io.github.thibaultbee.streampack.core.utils.extensions.getAutoWhiteBalanceModes
+import io.github.thibaultbee.streampack.core.utils.extensions.getCameraCharacteristics
+import io.github.thibaultbee.streampack.core.utils.extensions.getExposureMaxMeteringRegionsSupported
+import io.github.thibaultbee.streampack.core.utils.extensions.getExposureRange
+import io.github.thibaultbee.streampack.core.utils.extensions.getExposureStep
+import io.github.thibaultbee.streampack.core.utils.extensions.getFocusMaxMeteringRegionsSupported
+import io.github.thibaultbee.streampack.core.utils.extensions.getLensDistanceRange
+import io.github.thibaultbee.streampack.core.utils.extensions.getScalerMaxZoom
+import io.github.thibaultbee.streampack.core.utils.extensions.getSensitivityRange
+import io.github.thibaultbee.streampack.core.utils.extensions.getWhiteBalanceMeteringRegionsSupported
+import io.github.thibaultbee.streampack.core.utils.extensions.getZoomRatioRange
+import io.github.thibaultbee.streampack.core.utils.extensions.isFlashAvailable
+import io.github.thibaultbee.streampack.core.utils.extensions.isOpticalStabilizationAvailable
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledFuture
 import java.util.concurrent.TimeUnit
