+++ core/src/main/java/io/github/thibaultbee/streampack/core/utils/extensions/ContextExtensionsForCamera.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.thibaultbee.streampack.core.utils
+package io.github.thibaultbee.streampack.core.utils.extensions
 
 import android.content.Context
 import android.graphics.ImageFormat
